server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
RestrictedPython==7.0
pydantic==2.6.3
ruff==0.3.2
google-genai==1.69.0
python-dotenv==1.0.1